docs: PRD as repo file; refactor: cache-busting asset versions, remove dead code
- PRD.md (v2.2) in the app folder is now the single source of truth;
  issue #6 repointed there. README slimmed to user-facing content only
  (usage, tests, tech) with a link to the PRD
- Cache busting for mobile reloads: every local asset URL — CSS link,
  script tag and every inter-module import — carries the same ?v=N;
  bump N everywhere on release. New e2e checks fail on unversioned
  URLs or diverging versions
- Removed unused code: storage.clearState, decomposeEntry export
- CLAUDE.md: PRD-file location, README/PRD separation, version-bump rule
