+++ jass-scoreboard/index.html
@@ -5,7 +5,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Jass Tafel — Schiefertafel Z/Z</title>
   <link rel="icon" href="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 100 100'%3E%3Crect width='100' height='100' rx='18' fill='%23232e28'/%3E%3Cpath d='M25 32h50L25 68h50' stroke='%23f2efe4' stroke-width='9' fill='none' stroke-linecap='round' stroke-linejoin='round'/%3E%3C/svg%3E" />
-  <link rel="stylesheet" href="css/styles.css" />
+  <!-- Cache busting: bump ?v= here AND in every js import on each release -->
+  <link rel="stylesheet" href="css/styles.css?v=3" />
 </head>
 <body>
   <!-- Win overlay -->
@@ -94,6 +95,6 @@
     </div>
   </div>
 
-  <script type="module" src="js/app.js"></script>
+  <script type="module" src="js/app.js?v=3"></script>
 </body>
 </html>
