+++ jass-scoreboard/js/app.js
@@ -1,9 +1,9 @@
 // app.js — application bootstrap
 
-import { initState } from "./state.js";
-import { loadState } from "./storage.js";
-import { bindUI } from "./ui.js";
-import { render } from "./renderer.js";
+import { initState } from "./state.js?v=3";
+import { loadState } from "./storage.js?v=3";
+import { bindUI } from "./ui.js?v=3";
+import { render } from "./renderer.js?v=3";
 
 function bootstrap() {
   initState(loadState());
